# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Map4Accessibility Backend",
        "description": "Map4Accessibility Backend",
        "termsOfService": "http://swagger.io/terms/",
        "contact": {
            "email": "apiteam@swagger.io"
        },
        "license": {
            "name": "Apache 2.0",
            "url": "http://www.apache.org/licenses/LICENSE-2.0.html"
        },
        "version": "0.0.2"
    },
    "servers": [
        {
            "url": "/api/v1"
        }
    ],
    "paths": {
...
        "/faqs": {
            "get": {
                "tags": [
                    "FAQ's"
                ],
                "summary": "Get list of FAQs",
                "description": "Returns list of FAQs",
                "operationId": "getFaqs",
                "parameters": [
                    {
                        "name": "Accept-Language",
                        "in": "header",
                        "description": "Set language parameter",
                        "required": true,
                        "schema": {
                            "type": "string",
                            "example": "en"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Successful operation",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/FAQs"
                                }
                            }
                        }
                    }
                }
            }
        },
        "/legal-text/{type}": {
            "get": {
                "tags": [
                    "Legal Texts"
                ],
                "summary": "Get a legal text",
                "description": "Returns a legal text",
                "operationId": "getLegalText",
                "parameters": [
                    {
                        "name": "Accept-Language",
                        "in": "header",
                        "description": "Set language parameter",
                        "required": true,
                        "schema": {
                            "type": "string",
                            "example": "en"
                        }
                    },
                    {
                        "$ref": "#/components/parameters/LegalText--type"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Successful operation",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/LegalText"
                                }
                            }
                        }
                    }
                }
            }
        },
        "/auth/place-evaluation": {
            "post": {
                "tags": [
                    "PlaceEvaluation"
                ],
                "summary": "Create Place Evaluation by AppUser AUTH TOKEN",
                "description": "Create Place Evaluation by AppUser AUTH TOKEN",
                "operationId": "placeEvaluationByAuthenticated",
                "requestBody": {
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/PlaceEvaluation"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "successful operation",
                        "headers": {
                            "X-Rate-Limit": {
                                "description": "calls per hour allowed by the user",
                                "schema": {
                                    "type": "integer",
                                    "format": "int32"
                                }
                            },
                            "X-Expires-After": {
                                "description": "date in UTC when token expires",
                                "schema": {
                                    "type": "string",
                                    "format": "datetime"
                                }
                            }
                        },
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Invalid username/password supplied"
                    },
                    "400": {
                        "description": "Internal error"
                    }
                }
            }
        },
        "/auth/place-evaluation/{placeEvaluationId}/media": {
            "post": {
                "tags": [
                    "PlaceEvaluation"
                ],
                "summary": "upload a media file for Place Evaluation ID by AppUser AUTH TOKEN",
                "description": "upload a media file for Place Evaluation ID by AppUser AUTH TOKEN",
                "operationId": "attachMediaPlaceEvaluationByAuthenticated",
                "parameters": [
                    {
                        "name": "placeEvaluationId",
                        "in": "path",
                        "required": true
                    }
                ],
                "requestBody": {
                    "content": {
                        "multipart/form-data": {
                            "schema": {
                                "properties": {
                                    "media": {
                                        "type": "string",
                                        "format": "binary"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "successful operation",
                        "headers": {
                            "X-Rate-Limit": {
                                "description": "calls per hour allowed by the user",
                                "schema": {
                                    "type": "integer",
                                    "format": "int32"
                                }
                            },
                            "X-Expires-After": {
                                "description": "date in UTC when token expires",
                                "schema": {
                                    "type": "string",
                                    "format": "datetime"
                                }
                            }
                        },
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Invalid username/password supplied"
                    },
                    "400": {
                        "description": "Internal error"
                    }
                }
            }
        }
    },
    "components": {
        "schemas": {
...
            "FAQ": {
                "title": "FAQ Object",
                "description": "FAQ model",
                "required": [],
                "properties": {
                    "id": {
                        "title": "id",
                        "description": "id of Book",
                        "type": "integer",
                        "readOnly": "true",
                        "example": "1"
                    },
                    "question": {
                        "title": "question",
                        "type": "string",
                        "example": "Is this a question?"
                    },
                    "answer": {
                        "title": "answer",
                        "type": "string",
                        "example": "This is an answer!"
                    },
                    "order": {
                        "title": "order",
                        "type": "integer",
                        "example": "1"
                    },
                    "locale": {
                        "title": "locale",
                        "type": "string",
                        "example": "en"
                    },
                    "created_at": {
                        "title": "created_at",
                        "type": "dateTime",
                        "readOnly": "true",
                        "example": "2022-07-04T02:41:42.336Z"
                    },
                    "updated_at": {
                        "title": "updated_at",
                        "type": "dateTime",
                        "readOnly": "true",
                        "example": "2022-07-04T02:41:42.336Z"
                    }
                },
                "type": "object"
            },
            "FAQs": {
                "title": "FAQ Response",
                "properties": {
                    "success": {
                        "title": "success",
                        "type": "boolean",
                        "readOnly": "true",
                        "example": "true"
                    },
                    "message": {
                        "title": "message",
                        "type": "string",
                        "readOnly": "true",
                        "example": "null"
                    },
                    "result": {
                        "title": "result",
                        "properties": {
                            "data": {
                                "title": "data",
                                "type": "array",
                                "items": {
                                    "$ref": "#/components/schemas/FAQ"
                                }
                            }
                        },
                        "type": "object"
                    }
                },
                "type": "object"
            },
            "LegalText": {
                "title": "Legal Text Object",
                "description": "Legal Text model",
                "required": [],
                "properties": {
                    "id": {
                        "title": "id",
                        "description": "id of Legal Text",
                        "type": "integer",
                        "readOnly": "true",
                        "example": "1"
                    },
                    "type": {
                        "title": "type",
                        "type": "string",
                        "example": "terms"
                    },
                    "description": {
                        "title": "description",
                        "type": "string",
                        "example": "This is a Legal Text!"
                    },
                    "locale": {
                        "title": "locale",
                        "type": "string",
                        "example": "en"
                    },
                    "created_at": {
                        "title": "created_at",
                        "type": "dateTime",
                        "readOnly": "true",
                        "example": "2022-07-04T02:41:42.336Z"
                    },
                    "updated_at": {
                        "title": "updated_at",
                        "type": "dateTime",
                        "readOnly": "true",
                        "example": "2022-07-04T02:41:42.336Z"
                    }
                },
                "type": "object"
            },
...
            "PlaceEvaluation": {
                "properties": {
                    "id": {
                        "title": "ID",
                        "description": "placeEvaluationId",
                        "format": "int64",
                        "readOnly": 1
                    },
                    "google_place_id": {
                        "title": "Google Place id",
                        "description": "Google Place id",
                        "format": "int64",
                        "example": ""
                    },
                    "name": {
                        "title": "Name",
                        "description": "Name Place"
                    },
                    "country": {
                        "title": "country",
                        "description": "country"
                    },
                    "latitude": {
                        "title": "Coord. Latitude",
                        "description": "Coord. Latitude",
                        "format": "decimal",
                        "example": ""
                    },
                    "longitude": {
                        "title": "Coord. Longitude",
                        "description": "Coord. Longitude",
                        "format": "decimal",
                        "example": ""
                    },
                    "thumb_direction": {
                        "title": "Thumb Direction",
                        "description": "Thumb Direction boolean (0- thumb_down, 1- thumb_up)",
                        "type": "integer",
                        "maximum": 1,
                        "minimum": 0
                    },
                    "comment": {
                        "title": "Comment",
                        "description": "Comment"
                    },
                    "questions_answers": {
                        "title": "Questions Answers JSON",
                        "description": "Questions Answers JSON",
                        "type": "object",
                        "example": []
                    },
                    "media_url": {
                        "title": "Media Url Cloudinary",
                        "description": "Media Url Cloudinary",
                        "readOnly": 1
                    }
                },
                "type": "object"
            }
        },
        "parameters": {
            "LegalText--type": {
                "name": "type",
                "in": "path",
                "description": "Type of Legal Text",
                "required": true,
                "schema": {
                    "type": "string",
                    "example": "terms"
                }
            }
        },
        "securitySchemes": {
            "api_key_security": {
                "type": "apiKey",
                "description": "A API_KEY 1 security validation",
                "name": "x-api-key",
                "in": "header"
            },
            "sanctum": {
                "type": "apiKey",
                "description": "A 2 security validation, enter token in format (Bearer <token>)",
                "name": "Authorization",
                "in": "header"
            }
        }
    },
...
    "security": [
        {
            "api_key_security": [
                "read",
                "write"
            ],
            "sanctum": [
                "read",
                "write"
            ]
        }
    ]
}